+++ KitCli.Tests/CliAppBuilderTests.cs
@@ -71,7 +71,6 @@ public class CliAppBuilderTests
     {
         public void Register(IServiceCollection services)
         {
-            services.AddCliAbstractions();
             services.AddCliInstructions();
             services.AddCommandAbstractions();